re-combined the two apps
