+++ NR Status/Info.plist
