+++ NR Status/NR StatusRelease.entitlements
