+++ NR Status/Summaries/ApplicationGridView.swift
