+++ NR Status/Summaries/ApplicationView.swift
