--- NRQL EditorTests/NRQL_EditorTests.swift
@@ -1,17 +1,0 @@
-//
-//  NRQL_EditorTests.swift
-//  NRQL EditorTests
-//
-//  Created by Eric Baur on 10/14/24.
-//
-
-import Testing
-@testable import NRQL_Editor
-
-struct NRQL_EditorTests {
-
-    @Test func example() async throws {
-        // Write your test here and use APIs like `#expect(...)` to check expected conditions.
-    }
-
-}
